changed parser to take intervals, fixed some typos

+++ Assets/Scripts/burstNote.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class burstNote : MonoBehaviour
 {
     public string text;
-    public int burstLength;
+    public float burstLength;
     private UIManager UI;
     void Start() {
         UI = GameObject.FindWithTag("Score").GetComponent<UIManager>();
